Fix compiler warnings
bet_graph.c:663:28: warning: cast from function call of type 'double' to
      non-matching type 'gint' (aka 'int') [-Wbad-function-cast]
            index = (gint) gog_axis_map_from_view (x_map, event->x);
                           ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
bet_graph.c:950:33: warning: cast from function call of type 'gdouble'
      (aka 'double') to non-matching type 'gint' (aka 'int')
      [-Wbad-function-cast]
    prefs->gap_spinner = (gint) gtk_spin_button_get_value ( spinbutton );
                                ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
2 warnings generated.

+++ src/bet_graph.c
@@ -660,7 +660,7 @@ static gboolean bet_graph_on_motion ( GtkWidget *event_box,
             GogAxisMap *x_map;
 
             x_map = gog_chart_map_get_axis_map (map, 0);
-            index = (gint) gog_axis_map_from_view (x_map, event->x);
+            index = gog_axis_map_from_view (x_map, event->x);
 
             buf = g_strdup_printf ( _("date %s : value %s"), self->tab_vue_libelle[index-1],
                         utils_real_get_string_with_currency_from_double (
@@ -947,7 +947,7 @@ static void bet_graph_gap_spinner_changed ( GtkSpinButton *spinbutton,
     struct_bet_graph_prefs *prefs;
 
     prefs = self->prefs;
-    prefs->gap_spinner = (gint) gtk_spin_button_get_value ( spinbutton );
+    prefs->gap_spinner = gtk_spin_button_get_value ( spinbutton );
 
     /* on met à jour le graph */
     bet_graph_update_graph ( self );
